requirements_railway.txt
# Requirements for Railway deployment
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-pptx>=0.6.23
PyPDF2>=3.0.1
pypdf>=3.17.1
pdf2image>=1.16.3
pillow>=10.0.0
python-docx>=1.1.0
openai>=1.3.0
anthropic>=0.7.0
aiofiles>=23.2.0
pydantic>=2.5.0
jinja2>=3.1.0
reportlab>=4.0.0
markdown>=3.5.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
openpyxl>=3.1.0
requests>=2.31.0
python-dotenv>=1.0.0
gunicorn>=20.1.0
